helper/schema: zero value of a set should be empty
